Baxter simulated on V-REP writes hello
Here are the files of a scene in V-REP with a modified version of the
Baxter model (I deleted the right arm, all the child scripts and made
everything static, non dynamic, etc). The pen has two separated dummy
objects and those are used with the IK to position the left arm. The
pen that is really writing the text is invisible and has the dummy
objects attached to it (it is possible to use the one Baxter is
holding, but then if the movement is too fast the IK cannot solve on
time and the trace becomes not continuum…). The word hello was
generated moving the Sphere inside the scene “1_writer.ttt” and
exporting the data generated inside the graph object to a CSV file (it
is necessary to adjust the buffer configuration of the graph).